+++ src/parser/block/extension/fig.ts
@@ -1,5 +1,5 @@
 import { ExtensionParser } from '../../block';
-import { union, sequence, some, block, line, validate, rewrite, convert } from '../../../combinator';
+import { union, sequence, some, block, line, validate, rewrite, close, convert } from '../../../combinator';
 import { contentline } from '../../source';
 import { figure } from './figure';
 import { segment as seg_label } from '../../inline/extension/label';
@@ -7,13 +7,13 @@ import { segment as seg_code } from '../codeblock';
 import { segment as seg_math } from '../mathblock';
 import { segment as seg_blockquote } from '../blockquote';
 import { segment as seg_table } from './table';
-import { segment as seg_placeholder } from './placeholder';
+import { segment as seg_placeholder, placeholder } from './placeholder';
 
 import FigParser = ExtensionParser.FigParser;
 
 export const segment: FigParser.SegmentParser = block(validate(['[$', '$'],
   sequence([
-    line(seg_label),
+    line(close(seg_label, /^.*/)),
     union([
       seg_code,
       seg_math,
@@ -30,4 +30,7 @@ export const fig: FigParser = block(rewrite(segment, convert(
       .reduce((max, fence) => fence > max ? fence : max, '~~') + '~';
     return `${fence}figure ${source}\n\n${fence}`;
   },
-  union([figure]))));
+  union([
+    figure,
+    placeholder,
+  ]))));